extract DashboardPage into its own file
main_page.dart was 548 lines because it co-located the DashboardPage,
StatusCard, and DashboardCard widgets with the navigation shell. They
share no state with MainPage, so move them to lib/pages/dashboard_page.dart
and let main_page.dart focus on the shell.

Also fixes a stray duplicate SizedBox in the status row and replaces the
old 'ignore: use_build_context_synchronously' on the Channels card with
a proper context.mounted guard.
